Organizando os diretórios da GUI
