+++ organizador-receitas/src/main/java/gui/Project.java
@@ -1,0 +1,258 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
+ */
+package gui;
+
+/**
+ *
+ * @author pedro
+ */
+public class Project extends javax.swing.JFrame {
+
+    /**
+     * Creates new form Home
+     */
+    public Project() {
+        initComponents();
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    // <editor-fold defaultstate="collapsed" desc="Generated
+    // Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        jComboBox1 = new javax.swing.JComboBox<>();
+        Home = new javax.swing.JPanel();
+        jScrollPane1 = new javax.swing.JScrollPane();
+        tabela = new javax.swing.JTable();
+        header = new javax.swing.JPanel();
+        tituloHeader = new javax.swing.JLabel();
+        sectionInteration = new javax.swing.JPanel();
+        pesquisa = new javax.swing.JTextField();
+        novo = new javax.swing.JButton();
+        filtros = new javax.swing.JButton();
+        jButton1 = new javax.swing.JButton();
+
+        jComboBox1.setModel(
+                new javax.swing.DefaultComboBoxModel<>(new String[] { "Item 1", "Item 2", "Item 3", "Item 4" }));
+
+        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
+        setTitle("Organizador de Receitas");
+        setBackground(new java.awt.Color(255, 255, 255));
+        setName("Home"); // NOI18N
+
+        Home.setBackground(new java.awt.Color(255, 255, 255));
+        Home.setMinimumSize(new java.awt.Dimension(200, 200));
+
+        tabela.setModel(new javax.swing.table.DefaultTableModel(
+                new Object[][] {
+                        { null, null, null },
+                        { null, null, null },
+                        { null, null, null },
+                        { null, null, null }
+                },
+                new String[] {
+                        "Titulo", "Tipo", "Ações"
+                }));
+        jScrollPane1.setViewportView(tabela);
+        if (tabela.getColumnModel().getColumnCount() > 0) {
+            tabela.getColumnModel().getColumn(2).setPreferredWidth(100);
+            tabela.getColumnModel().getColumn(2).setMaxWidth(100);
+        }
+
+        header.setBackground(new java.awt.Color(204, 0, 255));
+
+        tituloHeader.setBackground(new java.awt.Color(255, 255, 255));
+        tituloHeader.setFont(new java.awt.Font("Nimbus Roman", 0, 24)); // NOI18N
+        tituloHeader.setForeground(new java.awt.Color(255, 255, 255));
+        tituloHeader.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        tituloHeader.setText("Organizador de Receita");
+
+        javax.swing.GroupLayout headerLayout = new javax.swing.GroupLayout(header);
+        header.setLayout(headerLayout);
+        headerLayout.setHorizontalGroup(
+                headerLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, headerLayout.createSequentialGroup()
+                                .addGap(0, 0, Short.MAX_VALUE)
+                                .addComponent(tituloHeader, javax.swing.GroupLayout.PREFERRED_SIZE, 373,
+                                        javax.swing.GroupLayout.PREFERRED_SIZE)));
+        headerLayout.setVerticalGroup(
+                headerLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addComponent(tituloHeader, javax.swing.GroupLayout.DEFAULT_SIZE, 65, Short.MAX_VALUE));
+
+        sectionInteration.setBackground(new java.awt.Color(255, 255, 255));
+
+        pesquisa.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                pesquisaActionPerformed(evt);
+            }
+        });
+
+        novo.setText("Novo");
+        novo.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                novoActionPerformed(evt);
+            }
+        });
+
+        filtros.setIcon(new javax.swing.ImageIcon(getClass().getResource("/gui/images/filtro.png"))); // NOI18N
+        filtros.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                filtrosActionPerformed(evt);
+            }
+        });
+
+        jButton1.setBackground(new java.awt.Color(204, 0, 255));
+        jButton1.setIcon(
+                new javax.swing.ImageIcon(getClass().getResource("/gui/images/big-search-len.png"))); // NOI18N
+
+        javax.swing.GroupLayout sectionInterationLayout = new javax.swing.GroupLayout(sectionInteration);
+        sectionInteration.setLayout(sectionInterationLayout);
+        sectionInterationLayout.setHorizontalGroup(
+                sectionInterationLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(sectionInterationLayout.createSequentialGroup()
+                                .addGap(34, 34, 34)
+                                .addComponent(novo)
+                                .addGap(18, 18, 18)
+                                .addComponent(pesquisa)
+                                .addGap(0, 0, 0)
+                                .addComponent(filtros, javax.swing.GroupLayout.PREFERRED_SIZE, 33,
+                                        javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(0, 0, 0)
+                                .addComponent(jButton1, javax.swing.GroupLayout.PREFERRED_SIZE, 32,
+                                        javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(72, 72, 72)));
+        sectionInterationLayout.setVerticalGroup(
+                sectionInterationLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(sectionInterationLayout.createSequentialGroup()
+                                .addGroup(sectionInterationLayout
+                                        .createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                        .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, sectionInterationLayout
+                                                .createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                                                .addComponent(filtros, javax.swing.GroupLayout.PREFERRED_SIZE, 24,
+                                                        javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                .addGroup(javax.swing.GroupLayout.Alignment.LEADING,
+                                                        sectionInterationLayout.createSequentialGroup()
+                                                                .addGap(15, 15, 15)
+                                                                .addGroup(sectionInterationLayout.createParallelGroup(
+                                                                        javax.swing.GroupLayout.Alignment.BASELINE)
+                                                                        .addComponent(pesquisa,
+                                                                                javax.swing.GroupLayout.PREFERRED_SIZE,
+                                                                                javax.swing.GroupLayout.DEFAULT_SIZE,
+                                                                                javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                                        .addComponent(novo))))
+                                        .addComponent(jButton1, javax.swing.GroupLayout.Alignment.TRAILING,
+                                                javax.swing.GroupLayout.PREFERRED_SIZE, 24,
+                                                javax.swing.GroupLayout.PREFERRED_SIZE))
+                                .addGap(15, 15, 15)));
+
+        javax.swing.GroupLayout HomeLayout = new javax.swing.GroupLayout(Home);
+        Home.setLayout(HomeLayout);
+        HomeLayout.setHorizontalGroup(
+                HomeLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addComponent(header, javax.swing.GroupLayout.Alignment.TRAILING,
+                                javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE,
+                                Short.MAX_VALUE)
+                        .addComponent(sectionInteration, javax.swing.GroupLayout.DEFAULT_SIZE,
+                                javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 580, Short.MAX_VALUE));
+        HomeLayout.setVerticalGroup(
+                HomeLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, HomeLayout.createSequentialGroup()
+                                .addComponent(header, javax.swing.GroupLayout.PREFERRED_SIZE,
+                                        javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(0, 0, 0)
+                                .addComponent(sectionInteration, javax.swing.GroupLayout.PREFERRED_SIZE,
+                                        javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(0, 0, 0)
+                                .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 361,
+                                        Short.MAX_VALUE)));
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
+        getContentPane().setLayout(layout);
+        layout.setHorizontalGroup(
+                layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addComponent(Home, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE,
+                                javax.swing.GroupLayout.PREFERRED_SIZE));
+        layout.setVerticalGroup(
+                layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(layout.createSequentialGroup()
+                                .addComponent(Home, javax.swing.GroupLayout.PREFERRED_SIZE,
+                                        javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(0, 0, Short.MAX_VALUE)));
+
+        pack();
+        setLocationRelativeTo(null);
+    }// </editor-fold>//GEN-END:initComponents
+
+    private void novoActionPerformed(java.awt.event.ActionEvent evt) {// GEN-FIRST:event_novoActionPerformed
+        // TODO add your handling code here:
+    }// GEN-LAST:event_novoActionPerformed
+
+    private void pesquisaActionPerformed(java.awt.event.ActionEvent evt) {// GEN-FIRST:event_pesquisaActionPerformed
+        // TODO add your handling code here:
+    }// GEN-LAST:event_pesquisaActionPerformed
+
+    private void filtrosActionPerformed(java.awt.event.ActionEvent evt) {// GEN-FIRST:event_filtrosActionPerformed
+        // TODO add your handling code here:
+    }// GEN-LAST:event_filtrosActionPerformed
+
+    /**
+     * @param args the command line arguments
+     */
+    public static void main(String args[]) {
+        /* Set the Nimbus look and feel */
+        // <editor-fold defaultstate="collapsed" desc=" Look and feel setting code
+        // (optional) ">
+        /*
+         * If Nimbus (introduced in Java SE 6) is not available, stay with the default
+         * look and feel.
+         * For details see
+         * http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html
+         */
+        try {
+            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
+                if ("Nimbus".equals(info.getName())) {
+                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
+                    break;
+                }
+            }
+        } catch (ClassNotFoundException ex) {
+            java.util.logging.Logger.getLogger(Project.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (InstantiationException ex) {
+            java.util.logging.Logger.getLogger(Project.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (IllegalAccessException ex) {
+            java.util.logging.Logger.getLogger(Project.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
+            java.util.logging.Logger.getLogger(Project.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        }
+        // </editor-fold>
+        // </editor-fold>
+
+        /* Create and display the form */
+        java.awt.EventQueue.invokeLater(new Runnable() {
+            public void run() {
+                new Project().setVisible(true);
+            }
+        });
+    }
+
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JPanel Home;
+    private javax.swing.JButton filtros;
+    private javax.swing.JPanel header;
+    private javax.swing.JButton jButton1;
+    private javax.swing.JComboBox<String> jComboBox1;
+    private javax.swing.JScrollPane jScrollPane1;
+    private javax.swing.JButton novo;
+    private javax.swing.JTextField pesquisa;
+    private javax.swing.JPanel sectionInteration;
+    private javax.swing.JTable tabela;
+    private javax.swing.JLabel tituloHeader;
+    // End of variables declaration//GEN-END:variables
+}
